+++ src/stubs.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 
 use crossbeam::channel;
 
-use crate::client::transport::{GlobalResponseIterator, MessageBus, ResponseIterator};
+use crate::transport::{GlobalResponseIterator, MessageBus, ResponseIterator};
 use crate::messages::{RequestMessage, ResponseMessage};
 use crate::Error;
